refactor global activity feed to use common activity feed component
