+++ frontend/components/ActivityItem/_styles.scss
@@ -1,4 +1,4 @@
-.host-activity-item {
+.activity-item {
   display: grid; // Grid system is used to create variable dashed line lengths
   grid-template-columns: 16px 16px 8px 1fr;
   grid-template-rows: max-content;
@@ -61,7 +61,7 @@
       background-color: $ui-off-white;
       cursor: pointer;
 
-      .host-activity-item__details-actions {
+      .activity-item__details-actions {
         visibility: visible;
       }
     }
@@ -111,19 +111,19 @@
   }
 
   &:first-child {
-    .host-activity-item__avatar-upper-dash {
+    .activity-item__avatar-upper-dash {
       border-right: none;
     }
   }
 
   &:last-child {
-    .host-activity-item__avatar-lower-dash {
+    .activity-item__avatar-lower-dash {
       border-right: none;
     }
   }
 
   /**
-   * Starting here are the styles for the host activity item when it is the
+   * Starting here are the styles for the activity item when it is the
    * only activity that is being displayed (controlled by the `soloActivity prop`.
    * We switch from grid to flexbox since we don't need the dashed lines anymore.
    * we also dont show to actions on hover
@@ -135,15 +135,15 @@
     display: flex;
     gap: $pad-medium;
 
-    .host-activity-item__avatar-wrapper {
+    .activity-item__avatar-wrapper {
       display: block;
     }
 
-    .host-activity-item__avatar-lower-dash {
+    .activity-item__avatar-lower-dash {
       display: none;
     }
 
-    .host-activity-item__details-wrapper {
+    .activity-item__details-wrapper {
       display: block;
       padding: 0;
       margin-bottom: 0;
@@ -154,7 +154,7 @@
       }
     }
 
-    .host-activity-item__details-actions {
+    .activity-item__details-actions {
       display: none;
     }
   }
