+++ frontend/pages/DashboardPage/cards/ActivityFeed/GlobalActivityItem/_styles.scss
@@ -1,4 +1,4 @@
-.activity-item {
+.global-activity-item {
   display: grid; // Grid system is used to create variable dashed line lengths
   grid-template-columns: 16px 16px 1fr;
   grid-template-rows: 32px max-content;
