+++ frontend/pages/DashboardPage/cards/ActivityFeed/GlobalActivityItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./GlobalActivityItem";
