--- frontend/pages/hosts/details/cards/Activity/HostActivityItem/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./HostActivityItem";
